Refactor user list API hooks out of learning resources. Dead code commented out
